requirements.txt
pandas
matplotlib
seaborn
scikit-learn
pickle-mixin
python-dotenv
